Installed should be false. And...
Installed should be false. Because this is only meant to be true for plugins that are installed to the engine. This is set automatically when the plugin is installed via the marketplace.

I presume WhitlistPlatforms was included to meet marketplace regulations. I find it more pragmatic to instead BlacklistPlatforms None.